fastapi-backend/dininginfoapi/requirements.txt
azure-functions
fastapi
PyMySQL
python-dotenv
uvicorn

